refactor(Categories): remove duplicate files in categories route
